EEG_Lightning/requirements.txt
flake8
yapf
isort
yacs
gdown
tb-nightly
future
scipy
scikit-learn
pandas
numpy
pytorch_metric_learning
torch_optimizer
openpyxl
optuna
pytorch_lightning=1.4.1
braindecode
MOABB